requirements.txt
asgiref==3.7.2
Django==4.2.4
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-multiselectfield==0.1.11
Pillow==10.0.0
psycopg2==2.9.7
python-decouple==3.8
sqlparse==0.4.4
tzdata==2023.3
